refactor: move consts from hook to corresponding files gf-23
