+++ apps/frontend/src/libs/hooks/use-pagination/libs/constants/constants.ts
@@ -1,0 +1,2 @@
+export { DEFAULT_VALUES } from "./default-values.constant.js";
+export { FIRST_PAGE } from "./first-page.constant.js";
