+++ apps/frontend/src/libs/hooks/use-pagination/libs/constants/default-values.constant.ts
@@ -1,0 +1,6 @@
+const DEFAULT_VALUES = {
+	PAGE: 1,
+	PAGE_SIZE: 10,
+};
+
+export { DEFAULT_VALUES };
