+++ apps/frontend/src/libs/hooks/use-pagination/libs/enums/enums.ts
@@ -1,0 +1,1 @@
+export { QueryParameter } from "./query-parameter.enum.js";
