+++ apps/frontend/src/libs/hooks/use-pagination/libs/enums/query-parameter.enum.ts
@@ -1,0 +1,6 @@
+const QueryParameter = {
+	page: "page",
+	pageSize: "pageSize",
+} as const;
+
+export { QueryParameter };
