+++ apps/frontend/src/libs/hooks/use-pagination/libs/helpers/helpers.ts
@@ -1,0 +1,1 @@
+export { calculateTotalPages } from "./total-pages.helper.js";
