+++ apps/frontend/src/libs/hooks/use-pagination/libs/helpers/total-pages.helper.ts
@@ -1,0 +1,4 @@
+const calculateTotalPages = (pageSize: number, totalItems: number): number =>
+	Math.ceil(totalItems / pageSize);
+
+export { calculateTotalPages };
